+++ src/modelNames.js
@@ -27,35 +27,52 @@ const fixModelName = (name) => {
   return name;
 };
 
-const formatModelName = (name) => {
+/**
+ *
+ * @param {string} name
+ * @param {{ ignorePrefix?: boolean; ignoreSuffix?: boolean }} options
+ * @returns
+ */
+const formatModelName = (name, options) => {
+  const typePrefix = options && options.ignorePrefix ? "" : config.typePrefix;
+  const typeSuffix = options && options.ignoreSuffix ? "" : config.typeSuffix;
+  const hashKey = `${typePrefix}_${name}_${typeSuffix}`;
+
   if (typeof name !== "string") {
     logger.warn("wrong name of the model name", name);
     return name;
   }
 
   if (/^([A-Z_]{1,})$/g.test(name)) {
-    return name;
+    return _.compact([typePrefix, name, typeSuffix]).join("_");
   }
 
-  if (formattedModelNamesMap.has(name)) {
-    return formattedModelNamesMap.get(name);
+  if (formattedModelNamesMap.has(hashKey)) {
+    return formattedModelNamesMap.get(hashKey);
   }
 
   const fixedModelName = fixModelName(name);
 
   const formattedModelName = _.replace(
-    _.startCase(`${config.typePrefix}_${fixedModelName}_${config.typeSuffix}`),
+    _.startCase(`${typePrefix}_${fixedModelName}_${typeSuffix}`),
     /\s/g,
     "",
   );
   const modelName = config.hooks.onFormatTypeName(formattedModelName, name) || formattedModelName;
 
-  formattedModelNamesMap.set(name, modelName);
+  formattedModelNamesMap.set(hashKey, modelName);
 
   return modelName;
 };
 
+const formatEnumKey = (key) =>
+  formatModelName(key, {
+    ignorePrefix: true,
+    ignoreSuffix: true,
+  });
+
 module.exports = {
   formatModelName: formatModelName,
+  formatEnumKey: formatEnumKey,
   isValidName,
 };