+++ tests/generate.js
@@ -1,20 +1,9 @@
-const { resolve } = require("path");
-const createSchemasInfos = require("./helpers/createSchemaInfos");
-const { generateApi } = require("../src");
+const allSchemas = require("./allSchemas");
+const { generateApiForTest } = require("./helpers/generateApiForTest");
 
-const schemas = [
-  ...createSchemasInfos({
-    absolutePathToSchemas: resolve(__dirname, "./schemas/v2.0"),
-    absoluteOutputPath: resolve(__dirname, "./generated/v2.0"),
-  }),
-  ...createSchemasInfos({
-    absolutePathToSchemas: resolve(__dirname, "./schemas/v3.0"),
-    absoluteOutputPath: resolve(__dirname, "./generated/v3.0"),
-  }),
-];
-
-schemas.forEach(({ absolutePath, apiFileName, outputPath }) => {
-  generateApi({
+allSchemas.forEach(({ absolutePath, apiFileName, outputPath }) => {
+  generateApiForTest({
+    testName: "generate script",
     name: apiFileName,
     input: absolutePath,
     output: outputPath,