+++ tests/spec/modular/test.js
@@ -1,41 +1,27 @@
-const { generateApi } = require("../../../src");
+const { generateApiForTest } = require("../../helpers/generateApiForTest");
 const { resolve } = require("path");
 const validateGeneratedModule = require("../../helpers/validateGeneratedModule");
-const createSchemasInfos = require("../../helpers/createSchemaInfos");
+const createSchemaInfos = require("../../helpers/createSchemaInfos");
 
-const schemas = createSchemasInfos({ absolutePathToSchemas: resolve(__dirname, "./") });
+const schemas = createSchemaInfos({ absolutePathToSchemas: resolve(__dirname, "./") });
 
-schemas.forEach(({ absolutePath }) => {
-  generateApi({
+schemas.forEach(({ absolutePath, Exception }) => {
+  generateApiForTest({
+    testName: "--modular option test",
     silent: true,
     input: absolutePath,
     output: resolve(__dirname, "./"),
     modular: true,
     generateClient: true,
     generateRouteTypes: true,
-  })
-    .then(() => {
-      const diagnostics = [
-        ...validateGeneratedModule({
-          pathToFile: resolve(__dirname, `./data-contracts.ts`),
-        }),
-        ...validateGeneratedModule({
-          pathToFile: resolve(__dirname, `./http-client.ts`),
-        }),
-        ...validateGeneratedModule({
-          pathToFile: resolve(__dirname, `./Key`),
-        }),
-        ...validateGeneratedModule({
-          pathToFile: resolve(__dirname, `./Login`),
-        }),
-        ...validateGeneratedModule({
-          pathToFile: resolve(__dirname, `./Scope`),
-        }),
-      ];
-      if (diagnostics.length) throw "Failed";
-    })
-    .catch((e) => {
-      console.error("modular option test failed.");
-      throw e;
-    });
+  }).then(() => {
+    const diagnostics = [
+      ...validateGeneratedModule(resolve(__dirname, `./data-contracts.ts`)),
+      ...validateGeneratedModule(resolve(__dirname, `./http-client.ts`)),
+      ...validateGeneratedModule(resolve(__dirname, `./Key`)),
+      ...validateGeneratedModule(resolve(__dirname, `./Login`)),
+      ...validateGeneratedModule(resolve(__dirname, `./Scope`)),
+    ];
+    if (diagnostics.length) throw new Exception("Failed");
+  });
 });